+++ src/irvine/oeis/a167/A167368.java
@@ -21,6 +21,6 @@ public class A167368 extends AbstractSequence {
   @Override
   public Z next() {
     ++mN;
-    return Functions.SIGMA.z(Functions.FACTORIAL.z(Functions.MULTIFACTORIAL.z(mN).intValue()));
+    return Functions.SIGMA.z(Functions.FACTORIAL.z(Functions.MULTIFACTORIAL.z(mN)));
   }
 }
