+++ src/irvine/oeis/a178/A178803.java
@@ -14,7 +14,7 @@ public class A178803 extends SimpleTransformSequence {
   /** Construct the sequence. */
   public A178803() {
     super(1, new A036043(), v -> {
-      return Functions.FACTORIAL.z(v.intValue());
+      return Functions.FACTORIAL.z(v);
     });
     skip(1);
   }
